chore: move 2930 tests into main test file
